[Fixed] Droid component modules crafting floating point decimals - Mantis #8352

+++ MMOCoreORB/src/server/zone/objects/tangible/components/droid/DroidAutoRepairModuleDataComponent.h
@@ -17,7 +17,7 @@ namespace droid {
 class DroidAutoRepairModuleDataComponent : public BaseDroidModuleComponent {
 
 protected:
-	int autoRepairPower;
+	float autoRepairPower;
 	bool active;
 
 public:
@@ -35,7 +35,7 @@ public:
 	void onStore();
 	void updateCraftingValues(CraftingValues* values, bool firstUpdate);
 	bool isActive(){ return active; }
-	int getAutoRepairPower(){ return autoRepairPower; }
+	float getAutoRepairPower() { return autoRepairPower; }
 	virtual bool isStackable() { return true; }
 	virtual void addToStack(BaseDroidModuleComponent* other);
 	virtual void copy(BaseDroidModuleComponent* other);
